Various changes
Minor changes to SpriteAnimation.
Deleted unnecessary player images.
Player now has to sort all waste correctly to advance. 
Renamed room images.
+ Collision detection for player and walls
+ Wall loading from JSON-file

+++ semesterprojekt/src/main/java/dk/sdu/worldoftrash/game/domain/items/Item.java
@@ -41,7 +41,7 @@ public class Item {
         this.game = game;
     }
 
-    public void update(float delta) {}
+    public void update(double delta) {}
 
     public void render(GraphicsContext gc) {};
